Fix multilib PACBTI-M mapping
Follow up to [#467](https://github.com/ARM-software/LLVM-embedded-toolchain-for-Arm/pull/467).

This patch fixes the multilib mapping for PACBTI-M.

The PACBTI-M variant should also match to configurations where
the branch protection is restricted to only one of PAC or BTI, like in
`-mbranch-protection=bti` or `-mbranch-protection=pac-ret`. Before this
patch, the variant would only match if both features were requested at
the same time. This is safe to do because instructions from either
feature have no effect if that feature is disabled at runtime.

This change also replaces ocurrences of `-mbranch-protection=standard`
by `-mbranch-protection=pac-ret+bti` because the latter is more
explicit.