Fix multilib PACBTI mapping (#501)
Follow up to
[#467](https://github.com/ARM-software/LLVM-embedded-toolchain-for-Arm/pull/467).

This patch fixes the multilib mapping for `-mbranch-protection`.

Firstly, the multilib flag to be inserted should be
`-mbranch-protection=standard` as it is the flag expected by the
PACBTI-M multilib variants.

Secondly, the PACBTI-M variant should also match to configurations where
the branch protection is restricted to only one of PAC or BTI, like in
`-mbranch-protection=bti` or `-mbranch-protection=pac-ret`. Before this
patch, the variant would only match if both features were requested at
the same time. This is safe to do because instructions from either
feature have no effect if that feature is disabled at runtime.