Adjust clang's multilib flags test to cope with missing variant
The clang driver's `print-multi-selection-flags.c` test uses a
combination of command line options in one of its RUN lines for which we
don't have a valid library variant available.
Since the merging of PR #500, this combination of options now trigger an
error to inform the user of an unsupported setup, but this error also
causes the test to fail.
This adds a new patch file for llvm-project to adjust the test
accordingly.