Use FVP models to test v8.1-M libraries with PACBTI (#506)
QEMU doesn't currently support the M-profile PACBTI extension, so we
need to use FVP models to test those libraries. There are also other
targets not supported by QEMU, such as v8-R, which we will need these
for.

These models are available free of cost (though not open source) from
Arm, and the user must agree to a EULA before using them. I've added a
script `fvp/get_fvps.sh' to download and extract them into the `fvp'
directory.

These models are currently only used for the v8.1-M libraries with
PACBTI, but the script also downloads the AEMv8A and AEMv8R models for
future use. If the FVPs aren't installed, then the CMakeLists.txt will
skip any tests which need them. I don't think it's worth preserving the
ability to run these tests with QEMU, because we still have the
non-PACBTI library variants which are run with QEMU. The two models also
have different memory maps and require different scatter files, so if we
wanted the ability to run tests for one library on both then we'd need
to add a way to build multiple linker scripts.