Add additional patch folder for performance patches
This adds a second llvm-project patch directory, intended for
performance enhancement patches which can be kept optional to avoid
blocking builds if they become out of date with the main branch. A
new CMake option controls whether or not to apply the patches, and a
new script is included to manage the patching process to make it
easier to identify when there is a problem with a patch.