Generate multilib.yaml for compatible architecture versions and features (#517)
This expands multilib-fpus.py (renaming it) to add Match blocks to split
the -march options into two sets of options, which will be more useful
for libray matching:
* A -march= option for each minor version less than or equal to the
selected one, so that it will be possible to mark a library as requiring
one architecture version, and allowing it to be used for any later
version. This strips any feature modifiers off the option, so that they
can be matched independently of base architecture version.
* A -march option for every feature modifier which is enabled or
disabled. This uses the made-up architecture name "armvX", which isn't a
valid option, but that doesn't matter to multilib, and this allows
matching features independently of the base architecture version.