Disable assertions for libunwind
When building the embedded toolchain, we currently disable assertions
for libcxxabi, but assertions for libunwind are still enabled. This
causes some of the logging functionality from libunwind to be included
in the build, which in turn imposes a dependency on IO functionality
that might not be available on baremetal environemnts.

This patch disables assertions for libunwind using the
`-DLIBUNWIND_ENABLE_ASSERTIONS=OFF` option, removing the issue.