Add LLVM libc sample (#538)
This will be built into the LLVM libc overlay package. Sample installs
to samples/src/llvmlibc/baremetal-semihosting.

* Add llvmlibc-samples directory to contain LLVM libc specific samples.
* Add single baremetal-semihosting sample using LLVM libc.

Sample contains a number of calls to LLVM libc functionality:
* heap.
* printf.
* string library.
* math library.

A special linker script is used to set the bounds of the heap. This uses
different linker defined symbols to picolibc.

The crt0.S from the existing llvmlibc-support can't be used directly as
the RW and ZI from the linker script need initializing.