Remove armv4t/armv5t from check-all targets
The compiler-rt and libcxx tests for armv4t and armv5t variants
currently hang until a timeout is reached. Rather than disable the
tests, this patch removes them from the check-all targets. This
allows the tests to still be run while investigating the issue, but
does not interrupt overall testing of the toolchain.