Fix MinGW toolchain install with 'install-llvm-toolchain'
When MinGW LLVM Embedded Toolchain for Arm is installed using the
`install-llvm-toolchain` target, the MinGW-built toolchain executables
are not installed under the prefix because `llvm-toolchain-mingw`
component is added to the `LLVM_TOOLCHAIN_DISTRIBUTION_COMPONENTS_ALL`
variable, which is never referenced.

Fix this by enumerating `LLVM_TOOLCHAIN_DISTRIBUTION_COMPONENTS_ALL` as
intended.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>