Move CMake code for building multilib and variants to subprojects
This patch restructures the top-level CMake script in order to split
building the library variants into a separate CMake project. This
better encapsulates the code to build the variants, and untethers it
from the toolchain builds so that libraries can be developed and
tested without having the rebuild the entire toolchain.

Similarly, a new subproject has been created to build a multilib set
of library variants, using a selected C library. A simple JSON format
has been used to define the flags and variants, which are then built
and collected into the appropriate layout.