Disable FVP testing by default
Previously, FVP testing was skipped if a FVP install was not
available. However, this now throws a configuration error if testing
relies on FVPs. But if you specifically want to test with FVPs, it is
desirable to have the configuration error warn you with an if the
install cannot be found.

To resolve this, an additional option has been added to explicitly
enable FVP testing, defaulting to OFF. This will override any
settings from the variant JSON, so that the default case remains
skipping FVP testing.