Corrections to CMake changes to enable library sub-builds (#571)
A number of corrections to the recent CMake changes:

* Two scripts arguments were renamed during review, but the cmake
steps calling the script weren't updated.

* Our downstream test targets such as check-package-llvm-toolchain
target require the path to an external llvm-lit in order to run the tests.

* The variables to enable exceptions and RTTI are now cached and the names
are capitalized, but CMake arguments needed updating to reflect this.

* The PREBUILT_TARGET_LIBRARIES option is meant to disable the building of
all libraries, on the assumption that pre-built libraries will be
copied from another location. The functionality was accidentally removed,
but has now been returned.

* CMake automatically expands a string containing semi-colons to a
list, but the ExternalProject argument must also be a semi-colon
separated list. To get around this, convert the
LLVM_TOOLCHAIN_LIBRARY_VARIANTS string to a comma separated list, since
the LIST_SEPARATOR option will reinterpret this as a semi-colon list.

* Previously, FVP testing was skipped if a FVP install was not
available. However, this now throws a configuration error if testing
relies on FVPs. But if you specifically want to test with FVPs, it is
desirable to have the configuration error warn you with an if the
install cannot be found. To resolve this, an additional option has 
been added to explicitly enable FVP testing, defaulting to OFF. This
will override any settings from the variant JSON, so that the default
case remains skipping FVP testing.