Make fvp work for anything other than aarch64 when being set as the test machine (#574)
The current picolibc changes can't handle fvp being set as the test
machine for anything other than aarch64. If fvp isn't set, it defaults
to qemu, which works for now because the ARM fvp doesn't require
anything special.
However, when adding new aarch64 big-endian variants, one of the
picolibc patches includes test-machine as an option in the Meson build
scripts. Consequently, the library's CMake needs to set this option
using -Dtest-machine=${TEST_EXECUTOR}. This means that for the armv8.1m
library variant,we are explicitly passing the FPU, which can cause the
error fvp: requested test machine not found since fvps are only defined
for the aarch64 case.