Fixes to patch application. (#579)
Every `CMakeLists.txt` that includes `fetch_llvm.cmake` must also call
`find_package(Python3)`, because `fetch_llvm.cmake` tries to run
`patch_llvm.py` by prefixing it with the Python interpreter name. The
CMakeLists in `arm-multilib` wasn't looking for Python at all, and the
one in `arm-runtimes` was looking for it _after_ including
`fetch_llvm.cmake`.

Also, `fetch_picolibc.cmake` was not reliably looking in the correct
directory for its patches: it was using a path relative to
`CMAKE_CURRENT_SOURCE_DIR`, which varies depending which CMakeLists it's
invoked by. It should be relative to `CMAKE_CURRENT_LIST_DIR`, as the
corresponding llvm and newlib paths already are.