Use the same patching script and command for all repos
To simplify patching, the script used for the LLVM repo can be
expanded for use on all repos. This also updates the newlib patch to
use the same format as llvm-project and picolibc; using git's patch
format means the .patch file is compatible with both git am and git
apply, whereas a simple diff only works with git apply.

This also adds an option to control the patch method used by the
script, since using git am may be preferable to the default git
apply. The script already supports the option, so all that is
needed to pass down the selection from the CMake cache if present.

Running git --check together with --apply --3way does not generatean
error return code when the patches are valid but contain conflicts,
so this particular combination is now marked as incompatible with
the --restore_on_fail option, since this relies on --check generating
an error.