Add an option to disable QEMU testing
Testing with QEMU is implicitly enabled by default, but configuring
the project to run with tests requires QEMU. As testing with QEMU
is considered optional, it should be possible to disable tests when
QEMU is not available or testing is not needed.

Since there is already an option for controlling FVP testing, this
patch adds a similar option for QEMU, ENABLE_QEMU_TESTING, which
instead defaults to ON. This can then be turned off when QEMU is
not installed, so that configurations where QEMU is expected can
suitably detect and show an error when it is missing.