Use the same patching script and command for all repos (#584)
To simplify patching, the script used for the LLVM repo can be
expanded for use on all repos. This also updates the newlib patch to
use the same format as llvm-project and picolibc; using git's patch
format means the .patch file is compatible with both git am and git
apply, whereas a simple diff only works with git apply.

This also adds an option to control the patch method used by the
script, since using git am may be preferable to the default git
apply. The script already supports the option, so all that is
needed to pass down the selection from the CMake cache if present.

The --3way option has been removed as a default when using --apply,
since it can leave conflict markers in place which will not get
detected by the script or CMake and will lead to a compilation error.
Instead, an option to the script has been added to use --3way. Since
--check --apply --3way does not return an error code when the patch
is valid but contains conflicts, the --restore_on_fail is now marked
as incompatible with that combination.