Prevent arm-runtime CMake arguments persisting over loops
The additional_cmake_args variable is used to override the defaults
and disable testing so that it can be controlled by the higher-level
CMake project. But the variable is currently set only when the tests
should be disabled. If the tests do not need to be disabled, the
value is not updated and the version from the previous iteration of
the loop will be used, potentially incorrectly disabling the tests.

To fix this, clear the value on evry loop if not needed.