Fix build failures in aarch64a_[be_]soft_nofp variants.
Those two variants have `-fno-exceptions -fno-rtti` in the
configuration file that specifies options for building the library.
None of the other library variants is built with this flags, not even
the other variants intended for use in _applications_ without
exceptions and RTTI.

If you try to build libc++abi with those flags, it fails to build,
because `private_typeinfo.cpp` is unconditionally included in the
library and unconditionally uses `dynamic_cast`.

(And I assume that it's OK to use `dynamic_cast` in that part of
libc++abi if the application will never use RTTI, because then it
won't include that object file in the first place.)