Do not build AArch64a soft no_fp variants using llvm-libc (#592)
LLVM-libc does not currently support no-neon soft float on AArch64
targets. To prevent building these until support is available, this
patch adds an option to explicitly set which C libraries a variant can
be built with. If the option is absent, the default is to build with all
library options.