Add patch to fix volatile access bug (#595)
We have a customer affected by this bug, which was fixed after LLVM 19,
and we'd like to do a release of LLVM-ET 19 with the fix.

LLVM PR: https://github.com/llvm/llvm-project/pull/111679

Original commit message:

[DAGISel] Keep flags when converting FP load/store to integer (#111679)

This DAG combine replaces a floating-point load/store pair which has no
other uses with an integer one, but did not copy the memory operand
flags to the new instructions, resulting in it dropping the volatile
flag. This optimisation is still valid if one or both of the
instructions is volatile, so we can copy over the whole
MachineMemOperand to generate volatile integer loads and stores where
needed.