Build AArch32 base variants with strict alignment
This patch adds `-mno-unaligned-access` to the compile flags of the
AArch32 base variants.

Base variants are the set of libraries that provide the most compatible
functionality. As a consequence, they are built with strict alignment in
order to work nicely with client code that is compiled with or without
strict alignment.

The libraries for Armv7-a soft already had a variant with and without
strict alignment. Hence now just one remains, the one with strict
alignment.